[add]add discarded_foods controller
