+++ views/discarded_foods/_show_message.erb
