+++ views/discarded_foods/food_discard.erb
@@ -4,9 +4,9 @@
 
 <h1>廃棄した食材の登録</h1>
 
-<%= erb :'user_foods/_show_message' %>
+<%= erb :'discarded_foods/_show_message' %>
 
-<form action="/user_foods/food_discard" method="post">
+<form action="/discarded_foods/food_discard" method="post">
   <div>
     <table>
       <thead>
