+++ views/home.erb
@@ -31,7 +31,7 @@
       </div>
       <div class="register">
           <a href="/user_foods/food_upload" class="btn btn-primary">冷蔵庫に食材登録</a>
-          <a href="/user_foods/food_discard" class="btn btn-primary">廃棄した食材登録</a>
+          <a href="/discarded_foods/food_discard" class="btn btn-primary">廃棄した食材登録</a>
       </div>
     </div>
   </div>
